[pytorch] reorder tracer code in generated VariableTypes
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38308

This PR doesn't add any new functionality. The purpose of this PR
is to validate reordering tracing code in variable kernel doesn't break
anything (which is a prerequisite of stacked change of moving tracing
logic into a new dispatch backend).

And it will be easier to bisect in case it breaks something which is not
covered by tests.

Differential Revision: [D21570685](https://our.internmc.facebook.com/intern/diff/D21570685/)
ghstack-source-id: 104422566