Add dynamic_cast asserts to CPU Loops.
On CUDA, we currently support casting loops dynamically (i.e. when the argument or return types of the lamba don't match the dtypes of the TensorIterator).
On CPU, before this change we would essentially reinterpret_cast, now we internal assert.  We could add dynamic_casting support in the future on CPU.

ghstack-source-id: 59e8dcd20a7eb25de39c4b7c89fd5f85f66b62a3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38817